requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.2
arrow==1.2.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.4
cryptography==36.0.0
flaky==3.7.0
future==0.18.2
h11==0.12.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
loguru==0.5.3
lxml==4.6.4
outcome==1.1.0
packaging==21.3
pefile==2021.9.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyTelegramBotAPI==4.2.0
pytest==6.2.5
python-dateutil==2.8.2
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
telebot==0.0.4
times==0.7
tk==0.1.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webencodings==0.5.1
win32-setctime==1.0.3
wsproto==1.0.0